fix:修改加载渲染预设逻辑，适配sRGB
